% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCSJpackage.R
\name{search_datasets}
\alias{search_datasets}
\title{Search Dryad repository for datasets containing a keyword}
\usage{
search_datasets(keyword, max_results = 100)
}
\arguments{
\item{keyword}{A character vector of keywords or phrases to search for.}

\item{max_results}{Maximum number of results to return}
}
\value{
A tibble with dataset title, authors, publication year, and DOI link
}
\description{
Given a keyword or phrase, the \code{search_datasets()} function queries the Dryad API and returns metadata for matching datasets in the form of a tibble. Each row in the result includes the dataset title, authors, publication year, DOI, and methods section (if available).
}
\examples{
search_datasets(c("climate", "butterfly"), 50)
}
