% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCSJpackage.R
\name{search_method}
\alias{search_method}
\title{Search for datasets that use a specific method}
\usage{
search_method(result, keywords, match_all = FALSE)
}
\arguments{
\item{result}{A tibble returned from \code{search_datasets()}.}

\item{keywords}{Character vector of search keywords.}

\item{match_all}{Logical. If TRUE, all keywords must appear (AND logic). If FALSE (default), any keyword may appear (OR logic).}
}
\value{
A filtered tibble of matching datasets.
}
\description{
Given a search result tibble from \code{search_datasets()}, \code{search_method()} filters datasets whose methods OR abstract contain one or more target keywords.
}
\examples{
search_datasets("climate", 50) |> search_method(c("experiment", "heatwave"))
}
